git-svn-id: https://svn.apache.org/repos/asf/directory/shared/trunk@686916 13f79535-47bb-0310-9956-ffa450edef68

+++ ldap/src/main/java/org/apache/directory/shared/ldap/codec/search/controls/EntryChangeControlCodec.java
@@ -33,8 +33,8 @@ import org.apache.directory.shared.ldap.util.StringTools;
 
 /**
  * A response control that may be returned by Persistent Search entry responses.
- * It contains addition change information to descrive the exact change that
- * occured to an entry. The exact details of this control are covered in section
+ * It contains addition change information to describe the exact change that
+ * occurred to an entry. The exact details of this control are covered in section
  * 5 of this (yes) expired draft: <a
  * href="http://www3.ietf.org/proceedings/01aug/I-D/draft-ietf-ldapext-psearch-03.txt">
  * Persistent Search Draft v03</a> which is printed out below for convenience:
@@ -85,7 +85,7 @@ public class EntryChangeControlCodec extends AbstractAsn1Object
 
     private ChangeType changeType = ChangeType.ADD;
 
-    private int changeNumber = UNDEFINED_CHANGE_NUMBER;
+    private long changeNumber = UNDEFINED_CHANGE_NUMBER;
 
     /** The previous DN */
     private LdapDN previousDn = null;
@@ -105,6 +105,7 @@ public class EntryChangeControlCodec extends AbstractAsn1Object
         super();
     }
 
+    
     /**
      * Compute the EntryChangeControl length 
      * 
@@ -112,7 +113,7 @@ public class EntryChangeControlCodec extends AbstractAsn1Object
      *   | 
      *   +--> 0x0A 0x0(1-4) [1|2|4|8] (changeType) 
      *  [+--> 0x04 L2 previousDN] 
-     *  [+--> 0x02 0x0(1-4) [0..2^31-1] (changeNumber)]
+     *  [+--> 0x02 0x0(1-4) [0..2^63-1] (changeNumber)]
      */
     public int computeLength()
     {
@@ -160,10 +161,12 @@ public class EntryChangeControlCodec extends AbstractAsn1Object
         {
             Value.encode( bb, previousDnBytes );
         }
+        
         if ( changeNumber != UNDEFINED_CHANGE_NUMBER )
         {
             Value.encode( bb, changeNumber );
         }
+        
         return bb;
     }
 
@@ -216,13 +219,13 @@ public class EntryChangeControlCodec extends AbstractAsn1Object
     }
 
 
-    public int getChangeNumber()
+    public long getChangeNumber()
     {
         return changeNumber;
     }
 
 
-    public void setChangeNumber( int changeNumber )
+    public void setChangeNumber( long changeNumber )
     {
         this.changeNumber = changeNumber;
     }
